test(apidom-ls): fix tests after AsyncAPI message paylod parse update
